move some database table methods into a helper class
